# be/cmd/swag/docs/swagger.yaml
basePath: /
definitions:
...
  model.JwtInput:
    properties:
      data:
        type: string
      user_id:
        type: string
    type: object
  model.JwtValidateInput:
    properties:
      data:
        type: string
    type: object
...
  response.ErrResponseData:
    properties:
      code:
        description: Ma status code
        type: integer
      detail:
        description: Thong bao loi
      error:
        type: string
    type: object
  response.ResponseData:
    properties:
      code:
        description: Ma status code
        type: integer
      data:
        description: Du lieu duoc return
      message:
        description: Thong bao loi
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: lytranvinh.work@gmail.com
    name: TEAM V
    url: example.com/be/
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: example.com/be/
  title: API Documentation Chat App
  version: 1.0.0
paths:
...
  /api/v1/mservice/get-chats-user:
    get:
      consumes:
      - application/json
      description: Get list chat of user with user id
      parameters:
      - description: 'Authorization Bearer token microservice (Eg: Bearer 123456)'
        in: header
        name: Authorization
        required: true
        type: string
      - description: user ID
        in: query
        name: user_id
        required: true
        type: string
      - description: limit number of chat
        in: query
        name: limit
        required: true
        type: string
      - description: page number
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Hanlde get list chat of user
      tags:
      - Microservice
  /api/v1/mservice/get-user-in-chat:
    get:
      consumes:
      - application/json
      description: Get information user in chat with chat id
      parameters:
      - description: 'Authorization Bearer token microservice (Eg: Bearer 123456)'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Chat ID
        in: query
        name: chat_id
        required: true
        type: string
      - description: limit number of chat
        in: query
        name: limit
        required: true
        type: string
      - description: page number
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Hanlde get infomation user in chat
      tags:
      - Microservice
  /api/v1/token/create_refresh_token:
    post:
      consumes:
      - application/json
      description: Testing create a new refresh JWT token
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.JwtInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Create a new JWT token test
      tags:
      - Token
  /api/v1/token/create_token:
    post:
      consumes:
      - application/json
      description: Testing create a new JWT token
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.JwtInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Create a new JWT token test
      tags:
      - Token
  /api/v1/token/valid_token:
    post:
      consumes:
      - application/json
      description: Validate token for access token and refresh token
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.JwtValidateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Validate token for access token and refresh token
      tags:
      - Token
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"